Show favorites list for user upon refresh
